avicii room 1.2 ändring av scenetransiction manager.
scene 2 är klar

+++ Avicii The Last Concert/Assets/Jesper/Scripts/GoToNextSceneTrigger.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 public class GoToNextSceneTrigger : SceneTransitionManager
 {
     [SerializeField]
-    private float m_Timer;
+    private int m_Timer;
 
     private void OnTriggerEnter(Collider other)
     {
